+++ .eslintrc
@@ -6,7 +6,6 @@
   },
   "extends": [
     "@budproj/eslint-config-base-ts",
-    "@budproj/eslint-config-prettier-ts",
     "@budproj/eslint-config-intl",
     "@budproj/eslint-config-react"
   ],
