+++ package.json
@@ -4,11 +4,13 @@
   "description": "Web interface for Bud's execution mode",
   "main": "src/index.js",
   "scripts": {
-    "start": "cross-env NODE_ENV=production node dist/index.js",
-    "build": "NODE_ENV=production next build && tsc --project ./tsconfig.server.json",
+    "start": "NODE_ENV=production node ./dist/index.js",
+    "build": "NODE_ENV=production run-s build:app build:server",
+    "build:app": "next build",
+    "build:server": "tsc --project ./tsconfig.server.json",
     "start:dev": "run-s intl:refresh server:watch",
     "server:watch": "nodemon",
-    "server:launch": "ts-node --project tsconfig.server.json ./src/index.ts",
+    "server:launch": "ts-node --project tsconfig.server.json ./src/server/index.ts",
     "compile": "run-s compile:intl compile:typescript",
     "compile:intl": "run-p intl:compile:ptBR intl:compile:enUS",
     "compile:typescript": "tsc",
@@ -19,6 +21,7 @@
     "lint:fix": "npm run lint -- --fix",
     "lint:timing": "TIMING=1 eslint . --ext ts,tsx",
     "lint:messages": "eslint **/messages.ts",
+    "format:and:lint": "run-s format lint:fix",
     "intl:generate:id": "echo '\u001b[36m➤ \u001b[90mGenerating react-intl message ids...\u001b[39m' && npm run lint:messages",
     "intl:refresh": "run-s intl:generate:id intl:extract",
     "intl:extract": "formatjs extract './src/**/*.ts(x)?' --out-file lang/pt-BR.json --id-interpolation-pattern '[sha512:contenthash:base64:6]'",
@@ -45,7 +48,6 @@
     "@babel/preset-react": "^7.12.7",
     "@budproj/eslint-config-base-ts": "^1.0.0",
     "@budproj/eslint-config-intl": "^1.0.0",
-    "@budproj/eslint-config-prettier-ts": "^1.0.0",
     "@budproj/eslint-config-react": "^1.0.0",
     "@budproj/prettier-config": "^1.0.0",
     "@formatjs/cli": "^2.13.5",
@@ -78,12 +80,13 @@
     "browser-env": "^3.3.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.5",
-    "eslint": "^7.13.0",
+    "eslint": "^7.14.0",
     "faker": "^5.1.0",
+    "husky": "^4.3.5",
     "jest": "^26.6.3",
     "jest-react-hooks-shallow": "^1.4.2",
     "jest-transform-graphql": "^2.1.0",
-    "lint-staged": "^10.5.1",
+    "lint-staged": "^10.5.3",
     "miragejs": "^0.1.41",
     "node-fetch": "^2.6.1",
     "nodemon": "^2.0.6",
@@ -104,7 +107,6 @@
     "@types/react-input-mask": "^3.0.0",
     "accepts": "^1.3.7",
     "chalk": "^4.1.0",
-    "cross-env": "^7.0.3",
     "crypto": "^1.0.1",
     "date-fns": "^2.16.1",
     "deepmerge": "^4.2.2",
@@ -112,7 +114,6 @@
     "formik": "^2.2.5",
     "framer-motion": "^2.9.4",
     "glob": "^7.1.6",
-    "husky": "^4.3.0",
     "koa": "^2.13.0",
     "lodash": "^4.17.20",
     "logform": "^2.2.0",
