--- src/index.ts
@@ -1,3 +1,0 @@
-import server from './server'
-
-server.launch() // eslint-disable-line @typescript-eslint/no-floating-promises
