+++ tsconfig.server.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["src/server/**/*.ts"],
+  "include": ["./src/server"],
   "compilerOptions": {
     "isolatedModules": false,
     "noEmit": false,
